# sandbox-api/docs/openapi.yml
openapi: 3.0.0
info:
  contact: {}
  description: API for manipulating filesystem, processes and network.
  title: Sandbox API
  version: 0.0.1
paths:
...
  "/network/process/{pid}/monitor":
    delete:
      description: Stop monitoring for new ports opened by a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Object containing PID and success message
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid process ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Stop monitoring ports for a process
      tags:
        - network
    post:
      description: Start monitoring for new ports opened by a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortMonitorRequest"
        description: Port monitor configuration
        required: true
      responses:
        "200":
          description: Object containing PID and success message
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Start monitoring ports for a process
      tags:
        - network
  "/network/process/{pid}/ports":
    get:
      description: Get a list of all open ports for a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Object containing PID and array of network.PortInfo
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid process ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Get open ports for a process
      tags:
        - network
  /network/tunnel:
    delete:
      description: "Stop the network tunnel and restore the original network configuration. WARNING: After disconnecting, the sandbox will lose all outbound internet connectivity (no egress). Inbound connections to the sandbox will still work. Use PUT /network/tunnel/config to re-establish the tunnel."
      responses:
        "200":
          description: Tunnel disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: No tunnel is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to stop tunnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Disconnect tunnel
      tags:
        - network
  /network/tunnel/config:
    put:
      description: Apply a new tunnel configuration on the fly. The existing tunnel is torn down and a new one is established. This endpoint is write-only; there is no corresponding GET to read the config back.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TunnelConfigRequest"
        description: Base64-encoded tunnel configuration
        required: true
      responses:
        "200":
          description: Configuration applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid tunnel configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to apply configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Update tunnel configuration
      tags:
        - network
...
servers:
  - url: https://sbx-{sandbox_id}-{workspace_id}.{region}.bl.run
    variables:
      sandbox_id:
        default: unknown
      workspace_id:
        default: unknown
      region:
        default: unknown
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      properties:
        error:
          example: Error message
          type: string
      required:
        - error
      type: object
...
    PortMonitorRequest:
      properties:
        callback:
          description: URL to call when a new port is detected
          example: http://localhost:3000/callback
          type: string
      type: object
...
    SuccessResponse:
      properties:
        message:
          example: File created successfully
          type: string
        path:
          example: /path/to/file
          type: string
      required:
        - message
      type: object
...
    TunnelConfigRequest:
      properties:
        config:
          description: Base64-encoded tunnel config JSON
          example: eyJsb2NhbF9pcCI6ICIxMC4wLjAuMS8zMiIsIC4uLn0=
          type: string
      required:
        - config
      type: object
...
security:
  - BearerAuth: []